requirements.txt
# Production dependencies for Vercel deployment (optimized for size)
# Remove Streamlit and heavy visualization libraries for serverless

# API & Server
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20

# PDF & Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# Data Processing & Analysis (lightweight)
numpy==1.26.4

# NLP & Text Processing (lightweight)
nltk==3.9.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3